Add "code" field to Trainee, Trainer, program. Trainee, class, program service

+++ src/main/java/htms/model/Module.java
@@ -18,7 +18,7 @@ public class Module {
     @GeneratedValue(strategy = GenerationType.UUID)
     private UUID id;
     private String description;
-    private Short moduleNo;
+    private Integer moduleNo;
     @URL
     private String link;
